# Performance improvements and timespan reporting