I added some logging to see how long things were taking, and noticed an opportunity to unblock what should have been (and now is) a set of concurrent node building tasks. Results below are from running the swag store project.

```
success open and validate gatsby-configs - 0.054s
success load plugins - 0.435s
success onPreInit - 0.022s
success initialize cache - 0.006s
success copy gatsby files - 0.018s
success onPreBootstrap - 0.009s
success createSchemaCustomization - 0.005s
Checked for operations in progress: 569.551ms
Initiated bulk operation query: 376.147ms
Completed bulk operation: 5619.528ms
Created nodes from bulk operation: 2378.194ms
createProductsOperation: 9807.659ms
success Checking for changed pages - 0.001s
success source and transform nodes - 9.839s
success building schema - 0.599s
info Total nodes: 494, SitePage nodes: 36 (use --verbose for breakdown)
success createPages - 0.039s
success Checking for changed pages - 0.000s
success createPagesStatefully - 0.043s
success update schema - 0.017s
success write out redirect data - 0.001s
success Build manifest and related icons - 0.086s
success onPostBootstrap - 0.089s
info bootstrap finished - 12.490s
success onPreExtractQueries - 0.001s
success extract queries from components - 0.301s
success write out requires - 0.007s
success run static queries - 3.510s - 3/3 0.85/s
success run page queries - 0.026s - 3/3 113.65/s
success Generating image thumbnails - 194.428s - 1196/1196 6.15/s
```